+++ vite.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
       "@assets": path.resolve(__dirname, "src/assets"),
       "@components": path.resolve(__dirname, "src/components"),
       "@pages": path.resolve(__dirname, "src/pages"),
+      "@types": path.resolve(__dirname, "src/types"),
     },
   },
   server: {
